requirements.txt
flask>=2.0
openpyxl>=3.0
python-docx>=0.8.11
mammoth>=1.4.19
gunicorn>=20.1.0
